Moved Collaborator classes into their own file.
Updated imports for this and the last commit.
